fix warnings

+++ src/computation/vector_random_pcim.cpp
@@ -22,7 +22,7 @@ void vector_random_pcim::vector_random_init() {
   other_probes.resize(n_total_probes);
   iota(other_probes.begin(), other_probes.end(), 0);
 
-  for (int i=0; i != lgn.size(); i++) {
+  for (long unsigned int i=0; i != lgn.size(); i++) {
     std::swap(other_probes[lgn[i]], other_probes[other_probes.size() - i - 1]);
   }
   other_probes.erase(other_probes.end() - lgn.size(), other_probes.end());
@@ -43,7 +43,7 @@ void vector_random_pcim::vector_random_tile_creation(int index) {
 
     frequency[value] += 1;
   }
-  for(int i=0; i<lgn.size(); i++) {
+  for(long unsigned int i=0; i<lgn.size(); i++) {
     auto it = random_it(tile_ptr->begin(), tile_ptr->end(), generator);
     tile_ptr->insert(it, lgn[i]);
     frequency[lgn[i]] += 1;
